refactor - #100 애플 소셜 로그인 request로 명시되어 있던 response값 이름 변경
