+++ src/main/java/sopt/org/hmh/global/auth/social/apple/fegin/ApplePublicKeyGenerator.java
@@ -11,23 +11,23 @@ import java.util.Map;
 import java.math.BigInteger;
 import sopt.org.hmh.global.auth.jwt.exception.JwtError;
 import sopt.org.hmh.global.auth.jwt.exception.JwtException;
-import sopt.org.hmh.global.auth.social.apple.request.ApplePublicKey;
-import sopt.org.hmh.global.auth.social.apple.request.ApplePublicKeys;
+import sopt.org.hmh.global.auth.social.apple.response.ApplePublicKeyResponse;
+import sopt.org.hmh.global.auth.social.apple.response.ApplePublicKeysResponse;
 
 @Component
 public class ApplePublicKeyGenerator {
-    public PublicKey generatePublicKeyWithApplePublicKeys(Map<String, String> headers, ApplePublicKeys applePublicKeys) {
-        ApplePublicKey applePublicKey = applePublicKeys
+    public PublicKey generatePublicKeyWithApplePublicKeys(Map<String, String> headers, ApplePublicKeysResponse applePublicKeysResponse) {
+        ApplePublicKeyResponse applePublicKeyResponse = applePublicKeysResponse
                 .getMatchesKey(headers.get("alg"), headers.get("kid"));
 
-        byte[] nBytes = Base64.getUrlDecoder().decode(applePublicKey.n());
-        byte[] eBytes = Base64.getUrlDecoder().decode(applePublicKey.e());
+        byte[] nBytes = Base64.getUrlDecoder().decode(applePublicKeyResponse.n());
+        byte[] eBytes = Base64.getUrlDecoder().decode(applePublicKeyResponse.e());
 
         RSAPublicKeySpec rsaPublicKeySpec = new RSAPublicKeySpec(
                 new BigInteger(1, nBytes), new BigInteger(1, eBytes));
 
         try {
-            KeyFactory keyFactory = KeyFactory.getInstance(applePublicKey.kty());
+            KeyFactory keyFactory = KeyFactory.getInstance(applePublicKeyResponse.kty());
             return keyFactory.generatePublic(rsaPublicKeySpec);
         } catch (NoSuchAlgorithmException | InvalidKeySpecException exception) {
             throw new JwtException(JwtError.UNABLE_TO_CREATE_APPLE_PUBLIC_KEY);
