--- src/main/java/sopt/org/hmh/global/auth/social/apple/request/ApplePublicKeys.java
@@ -1,18 +1,0 @@
-package sopt.org.hmh.global.auth.social.apple.request;
-
-import java.util.List;
-import lombok.Getter;
-import sopt.org.hmh.global.auth.jwt.exception.JwtError;
-import sopt.org.hmh.global.auth.jwt.exception.JwtException;
-
-@Getter
-public class ApplePublicKeys {
-    private List<ApplePublicKey> keys;
-
-    public ApplePublicKey getMatchesKey(String alg, String kid) {
-        return keys.stream()
-                .filter(applePublicKey -> applePublicKey.alg().equals(alg) && applePublicKey.kid().equals(kid))
-                .findFirst()
-                .orElseThrow(() -> new JwtException(JwtError.INVALID_IDENTITY_TOKEN));
-    }
-}
